+++ tools/daemon/command.do.pl
@@ -42,7 +42,6 @@ $| = 1;
 
 use TTP::Daemon;
 use TTP::Finder;
-my $running = $ep->runner();
 
 my $defaults = {
 	help => 'no',
@@ -68,7 +67,7 @@ my $opt_timeout = $defaults->{timeout};
 
 sub doSend {
 	# in dummy mode, just simulate and output the acknowledge
-	if( $running->dummy()){
+	if( $ep->runner()->dummy()){
 		msgDummy( "OK" );
 
 	# connect, triggering an error if the daemon is not active
@@ -151,18 +150,18 @@ if( !GetOptions(
 	"command=s"			=> \$opt_command,,
 	"timeout=i"			=> \$opt_timeout )){
 
-		msgOut( "try '".$running->command()." ".$running->verb()." --help' to get full usage syntax" );
+		msgOut( "try '".$ep->runner()->command()." ".$ep->runner()->verb()." --help' to get full usage syntax" );
 		TTP::exit( 1 );
 }
 
-if( $running->help()){
-	$running->verbHelp( $defaults );
+if( $ep->runner()->help()){
+	$ep->runner()->verbHelp( $defaults );
 	TTP::exit();
 }
 
-msgVerbose( "got colored='".( $running->colored() ? 'true':'false' )."'" );
-msgVerbose( "got dummy='".( $running->dummy() ? 'true':'false' )."'" );
-msgVerbose( "got verbose='".( $running->verbose() ? 'true':'false' )."'" );
+msgVerbose( "got colored='".( $ep->runner()->colored() ? 'true':'false' )."'" );
+msgVerbose( "got dummy='".( $ep->runner()->dummy() ? 'true':'false' )."'" );
+msgVerbose( "got verbose='".( $ep->runner()->verbose() ? 'true':'false' )."'" );
 msgVerbose( "got json='$opt_json'" );
 msgVerbose( "got name='$opt_name'" );
 msgVerbose( "got port='$opt_port'" );