+++ tools/dbms/vars.do.pl
@@ -36,7 +36,6 @@ use utf8;
 use warnings;
 
 use TTP::Service;
-my $running = $ep->runner();
 
 my $defaults = {
 	help => 'no',
@@ -114,18 +113,18 @@ if( !GetOptions(
 	"archivesDir!"		=> \$opt_archivesDir,
 	"service=s"			=> \$opt_service )){
 
-		msgOut( "try '".$running->command()." ".$running->verb()." --help' to get full usage syntax" );
+		msgOut( "try '".$ep->runner()->command()." ".$ep->runner()->verb()." --help' to get full usage syntax" );
 		TTP::exit( 1 );
 }
 
-if( $running->help()){
-	$running->verbHelp( $defaults );
+if( $ep->runner()->help()){
+	$ep->runner()->verbHelp( $defaults );
 	TTP::exit();
 }
 
-msgVerbose( "got colored='".( $running->colored() ? 'true':'false' )."'" );
-msgVerbose( "got dummy='".( $running->dummy() ? 'true':'false' )."'" );
-msgVerbose( "got verbose='".( $running->verbose() ? 'true':'false' )."'" );
+msgVerbose( "got colored='".( $ep->runner()->colored() ? 'true':'false' )."'" );
+msgVerbose( "got dummy='".( $ep->runner()->dummy() ? 'true':'false' )."'" );
+msgVerbose( "got verbose='".( $ep->runner()->verbose() ? 'true':'false' )."'" );
 msgVerbose( "got backupsRoot='".( $opt_backupsRoot ? 'true':'false' )."'" );
 msgVerbose( "got backupsDir='".( $opt_backupsDir ? 'true':'false' )."'" );
 msgVerbose( "got archivesRoot='".( $opt_archivesRoot ? 'true':'false' )."'" );