+++ tools/mqtt/publish.do.pl
@@ -33,7 +33,6 @@ use utf8;
 use warnings;
 
 use TTP::MQTT;
-my $running = $ep->runner();
 
 my $defaults = {
 	help => 'no',
@@ -56,7 +55,7 @@ sub doPublish {
 	msgOut( "publishing '$opt_topic [$opt_payload]'..." );
 	my $result = false;
 
-	if( $running->dummy()){
+	if( $ep->runner()->dummy()){
 		msgDummy( "considering publication successful" );
 		$result = true;
 	} else {
@@ -92,18 +91,18 @@ if( !GetOptions(
 	"payload=s"			=> \$opt_payload,
 	"retain!"			=> \$opt_retain	)){
 
-		msgOut( "try '".$running->command()." ".$running->verb()." --help' to get full usage syntax" );
+		msgOut( "try '".$ep->runner()->command()." ".$ep->runner()->verb()." --help' to get full usage syntax" );
 		TTP::exit( 1 );
 }
 
-if( $running->help()){
-	$running->verbHelp( $defaults );
+if( $ep->runner()->help()){
+	$ep->runner()->verbHelp( $defaults );
 	TTP::exit();
 }
 
-msgVerbose( "got colored='".( $running->colored() ? 'true':'false' )."'" );
-msgVerbose( "got dummy='".( $running->dummy() ? 'true':'false' )."'" );
-msgVerbose( "got verbose='".( $running->verbose() ? 'true':'false' )."'" );
+msgVerbose( "got colored='".( $ep->runner()->colored() ? 'true':'false' )."'" );
+msgVerbose( "got dummy='".( $ep->runner()->dummy() ? 'true':'false' )."'" );
+msgVerbose( "got verbose='".( $ep->runner()->verbose() ? 'true':'false' )."'" );
 msgVerbose( "got topic='$opt_topic'" );
 msgVerbose( "got payload='$opt_payload'" );
 msgVerbose( "got retain='".( $opt_retain ? 'true':'false' )."'" );