+++ tools/ovh/list.do.pl
@@ -29,7 +29,6 @@ use utf8;
 use warnings;
 
 use TTP::Ovh;
-my $running = $ep->runner();
 
 my $defaults = {
 	help => 'no',
@@ -102,18 +101,18 @@ if( !GetOptions(
 	"verbose!"			=> \$ep->{run}{verbose},
 	"services!"			=> \$opt_services )){
 
-		msgOut( "try '".$running->command()." ".$running->verb()." --help' to get full usage syntax" );
+		msgOut( "try '".$ep->runner()->command()." ".$ep->runner()->verb()." --help' to get full usage syntax" );
 		TTP::exit( 1 );
 }
 
-if( $running->help()){
-	$running->verbHelp( $defaults );
+if( $ep->runner()->help()){
+	$ep->runner()->verbHelp( $defaults );
 	TTP::exit();
 }
 
-msgVerbose( "got colored='".( $running->colored() ? 'true':'false' )."'" );
-msgVerbose( "got dummy='".( $running->dummy() ? 'true':'false' )."'" );
-msgVerbose( "got verbose='".( $running->verbose() ? 'true':'false' )."'" );
+msgVerbose( "got colored='".( $ep->runner()->colored() ? 'true':'false' )."'" );
+msgVerbose( "got dummy='".( $ep->runner()->dummy() ? 'true':'false' )."'" );
+msgVerbose( "got verbose='".( $ep->runner()->verbose() ? 'true':'false' )."'" );
 msgVerbose( "got services='".( $opt_services ? 'true':'false' )."'" );
 
 if( !TTP::errs()){