+++ tools/ttp/movedirs.do.pl
@@ -43,7 +43,6 @@ use Config;
 use File::Spec;
 
 use TTP::Path;
-my $running = $ep->runner();
 
 my $defaults = {
 	help => 'no',
@@ -193,18 +192,18 @@ if( !GetOptions(
 	"keep=s"			=> \$opt_keep,
 	"makeDirExist!"		=> \$opt_makeDirExist )){
 
-		msgOut( "try '".$running->command()." ".$running->verb()." --help' to get full usage syntax" );
+		msgOut( "try '".$ep->runner()->command()." ".$ep->runner()->verb()." --help' to get full usage syntax" );
 		TTP::exit( 1 );
 }
 
-if( $running->help()){
-	$running->verbHelp( $defaults );
+if( $ep->runner()->help()){
+	$ep->runner()->verbHelp( $defaults );
 	TTP::exit();
 }
 
-msgVerbose( "got colored='".( $running->colored() ? 'true':'false' )."'" );
-msgVerbose( "got dummy='".( $running->dummy() ? 'true':'false' )."'" );
-msgVerbose( "got verbose='".( $running->verbose() ? 'true':'false' )."'" );
+msgVerbose( "got colored='".( $ep->runner()->colored() ? 'true':'false' )."'" );
+msgVerbose( "got dummy='".( $ep->runner()->dummy() ? 'true':'false' )."'" );
+msgVerbose( "got verbose='".( $ep->runner()->verbose() ? 'true':'false' )."'" );
 msgVerbose( "got sourcepath='$opt_sourcepath'" );
 msgVerbose( "got sourcecmd='$opt_sourcecmd'" );
 msgVerbose( "got targetpath='$opt_targetpath'" );