+++ tools/ttp/sizedir.do.pl
@@ -41,7 +41,6 @@ use File::Find;
 use TTP::Constants qw( :all );
 use TTP::Message qw( :all );
 use TTP::Metric;
-my $running = $ep->runner();
 
 my $defaults = {
 	help => 'no',
@@ -98,7 +97,7 @@ sub doComputeSize {
 		$path =~ s/\//_/g;
 		$path =~ s/\\/_/g;
 		my @labels = ( @opt_prepends,
-			"environment=".$ep->node()->environment(), "command=".$running->command(), "verb=".$running->verb(), "path=$path",
+			"environment=".$ep->node()->environment(), "command=".$ep->runner()->command(), "verb=".$ep->runner()->verb(), "path=$path",
 			@opt_appends );
 		TTP::Metric->new( $ep, {
 			name => 'dirs_count',
@@ -167,18 +166,18 @@ if( !GetOptions(
 	"prepend=s@"		=> \@opt_prepends,
 	"append=s@"			=> \@opt_appends )){
 
-		msgOut( "try '".$running->command()." ".$running->verb()." --help' to get full usage syntax" );
+		msgOut( "try '".$ep->runner()->command()." ".$ep->runner()->verb()." --help' to get full usage syntax" );
 		TTP::exit( 1 );
 }
 
-if( $running->help()){
-	$running->verbHelp( $defaults );
+if( $ep->runner()->help()){
+	$ep->runner()->verbHelp( $defaults );
 	TTP::exit();
 }
 
-msgVerbose( "got colored='".( $running->colored() ? 'true':'false' )."'" );
-msgVerbose( "got dummy='".( $running->dummy() ? 'true':'false' )."'" );
-msgVerbose( "got verbose='".( $running->verbose() ? 'true':'false' )."'" );
+msgVerbose( "got colored='".( $ep->runner()->colored() ? 'true':'false' )."'" );
+msgVerbose( "got dummy='".( $ep->runner()->dummy() ? 'true':'false' )."'" );
+msgVerbose( "got verbose='".( $ep->runner()->verbose() ? 'true':'false' )."'" );
 msgVerbose( "got dirpath='$opt_dirpath'" );
 msgVerbose( "got dircmd='$opt_dircmd'" );
 msgVerbose( "got mqtt='".( $opt_mqtt ? 'true':'false' )."'" );