+++ tools/ttp/vars.do.pl
@@ -36,8 +36,6 @@ use strict;
 use utf8;
 use warnings;
 
-my $running = $ep->runner();
-
 my $defaults = {
 	help => 'no',
 	colored => 'no',
@@ -163,18 +161,18 @@ if( !GetOptions(
 	"alertsDir!"		=> \$opt_alertsDir,
 	"key=s"				=> \@opt_keys )){
 
-		msgOut( "try '".$running->command()." ".$running->verb()." --help' to get full usage syntax" );
+		msgOut( "try '".$ep->runner()->command()." ".$ep->runner()->verb()." --help' to get full usage syntax" );
 		TTP::exit( 1 );
 }
 
-if( $running->help()){
-	$running->verbHelp( $defaults );
+if( $ep->runner()->help()){
+	$ep->runner()->verbHelp( $defaults );
 	TTP::exit();
 }
 
-msgVerbose( "got colored='".( $running->colored() ? 'true':'false' )."'" );
-msgVerbose( "got dummy='".( $running->dummy() ? 'true':'false' )."'" );
-msgVerbose( "got verbose='".( $running->verbose() ? 'true':'false' )."'" );
+msgVerbose( "got colored='".( $ep->runner()->colored() ? 'true':'false' )."'" );
+msgVerbose( "got dummy='".( $ep->runner()->dummy() ? 'true':'false' )."'" );
+msgVerbose( "got verbose='".( $ep->runner()->verbose() ? 'true':'false' )."'" );
 msgVerbose( "got siteSpec='".( $opt_siteSpec ? 'true':'false' )."'" );
 msgVerbose( "got nodeRoot='".( $opt_nodeRoot ? 'true':'false' )."'" );
 msgVerbose( "got nodesDirs='".( $opt_nodesDirs ? 'true':'false' )."'" );